+++ src/main/java/org/corpus_tools/cffmaven/AbstractCffMojo.java
@@ -58,6 +58,7 @@ import org.ehcache.config.units.MemoryUnit;
 import org.json.JSONArray;
 import org.json.JSONObject;
 import org.snakeyaml.engine.v2.api.Load;
+import org.snakeyaml.engine.v2.api.LoadSettings;
 
 /**
  * Common functionality of the different CFF Mojos.
@@ -110,25 +111,53 @@ public abstract class AbstractCffMojo extends AbstractMojo {
   @Parameter
   protected List<String> ignoredArtifacts;
 
+  @Parameter
+  private List<TemplateConfiguration> referenceTemplates;
+
   private List<Pattern> ignoredPatterns;
 
   private Cache<String, RemoteLicenseInformation> remoteLicenseCache;
   private PersistentCacheManager cacheManager;
+  private Map<Pattern, File> templatePatterns;
+
+  private Load yamlLoad = new Load(LoadSettings.builder().build());
 
   protected Map<String, Object> createReference(Artifact artifact,
       ProjectBuildingRequest projectBuildingRequest) throws ProjectBuildingException {
-    LinkedHashMap<String, Object> reference = new LinkedHashMap<>();
-    reference.put("type", "software");
-    reference.put(TITLE, artifact.getArtifactId());
-    reference.put(VERSION, artifact.getVersion());
+    Map<String, Object> templateRef = null;
 
-    if (P2_PLUGIN_GROUP_ID.matcher(artifact.getGroupId()).matches()) {
-      createReferenceFromP2(reference, artifact, projectBuildingRequest);
-    } else {
-      createReferenceFromMavenArtifact(reference, artifact, projectBuildingRequest);
+    for (Map.Entry<Pattern, File> entry : getTemplatePatterns().entrySet()) {
+      getLog().debug(
+          "Testing artifact " + artifact.toString() + " with pattern " + entry.getKey().pattern());
+      if (entry.getKey().matcher(artifact.toString()).matches()) {
+        try {
+          getLog().info("Adding reference " + artifact.toString() + " from template "
+              + entry.getValue().getPath());
+          templateRef = createReferenceFromTemplate(artifact, projectBuildingRequest,
+              entry.getValue(), yamlLoad);
+          break;
+        } catch (IOException e) {
+          getLog().error("Could create reference from template " + entry.getValue().getPath(), e);
+        }
+      }
     }
 
-    return reference;
+    if (templateRef == null) {
+      // no pattern matched, use P2 or Maven information
+      LinkedHashMap<String, Object> reference = new LinkedHashMap<>();
+      reference.put("type", "software");
+      reference.put(TITLE, artifact.getArtifactId());
+      reference.put(VERSION, artifact.getVersion());
+
+      if (P2_PLUGIN_GROUP_ID.matcher(artifact.getGroupId()).matches()) {
+        createReferenceFromP2(reference, artifact, projectBuildingRequest);
+      } else {
+        createReferenceFromMavenArtifact(reference, artifact, projectBuildingRequest);
+      }
+      return reference;
+    } else {
+      return templateRef;
+    }
   }
 
 
@@ -501,6 +530,19 @@ public abstract class AbstractCffMojo extends AbstractMojo {
       }
     }
     return false;
+  }
+
+  protected Map<Pattern, File> getTemplatePatterns() {
+    if (templatePatterns == null) {
+      templatePatterns = new LinkedHashMap<Pattern, File>();
+      if (referenceTemplates != null) {
+        for (TemplateConfiguration config : referenceTemplates) {
+          Pattern p = Pattern.compile(config.getPattern().toString());
+          templatePatterns.put(p, config.getTemplate());
+        }
+      }
+    }
+    return templatePatterns;
   }