+++ CHANGELOG.md
@@ -6,6 +6,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed 
+
+- The `third-party-folder` goal now also actually respects the
+  `referenceTemplates` parameter. In case you have these configured, this also
+  speeds the lookup for P2 artifacts and makes sure the information in
+  `CITATION.cff` and the `THIRD-PARTY` folder are synchronized. This was already
+  described in the documentation, but not actually implemented.
+
+### Added
+
+- Added cache for requests to external servers (like clearlydefined.io or the
+  SPDX license server).
+
+
 ## [0.7.0] - 2023-03-03
 
 ### Fixed